+++ AudioConverterFileConverter/AppDelegate.cs
@@ -13,10 +13,10 @@
 // distribute, sublicense, and/or sell copies of the Software, and to
 // permit persons to whom the Software is furnished to do so, subject to
 // the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be
 // included in all copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 // EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -30,9 +30,9 @@ using System;
 
 using System.Diagnostics;
 using System.Threading;
-using MonoTouch.AVFoundation;
-using MonoTouch.Foundation;
-using MonoTouch.UIKit;
+using AVFoundation;
+using Foundation;
+using UIKit;
 
 namespace AudioConverterFileConverter
 {
@@ -65,7 +65,7 @@ namespace AudioConverterFileConverter
         private AVAudioSession session;
 
         // class-level declarations
-		
+
         public override UIWindow Window
         {
             get;
@@ -94,7 +94,7 @@ namespace AudioConverterFileConverter
                 ThreadStateEndInterruption();
             };
 
-            int routes = session.InputNumberOfChannels;
+			int routes = (int)session.InputNumberOfChannels;
 
             // our default category -- we change this for conversion and playback appropriately
             try
@@ -105,19 +105,19 @@ namespace AudioConverterFileConverter
             {
                 Debug.Print("ERROR: Cannot change audio session category");
             }
-                
+
             session.InputAvailabilityChanged += (object sender, AVStatusEventArgs e) =>
             {
                 var sess = sender as AVAudioSession;
                 Debug.Print("Old route: {0}", routes);
                 Debug.Print("New route: {0}", sess.InputNumberOfChannels);
-                routes = sess.InputNumberOfChannels;
+				routes = (int)sess.InputNumberOfChannels;
             };
 
             session.SetActive(true);
             return true;
         }
-		
+
         // This method is invoked when the application is about to move from active to inactive state.
         // OpenGL applications should use this method to pause.
         public override void OnResignActivation(UIApplication application)
@@ -139,7 +139,7 @@ namespace AudioConverterFileConverter
         {
         }
 
-        /// This method is called when the application is about to terminate. Save data, if needed. 
+        /// This method is called when the application is about to terminate. Save data, if needed.
         public override void WillTerminate(UIApplication application)
         {
         }
@@ -163,7 +163,7 @@ namespace AudioConverterFileConverter
                 if (State == State.Paused)
                 {
                     State = State.Running;
-				
+
                     StateChanged.Set();
                 }
             }
@@ -186,13 +186,13 @@ namespace AudioConverterFileConverter
             lock (StateLock)
             {
                 Debug.Assert(State != State.Done);
-			
+
                 while (State == State.Paused)
                 {
                     StateChanged.WaitOne();
                     wasInterrupted = true;
                 }
-			
+
                 // we must be running or something bad has happened
                 Debug.Assert(State == State.Running);
             }
